backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
python-magic==0.4.27
aiofiles==23.2.1
email-validator==2.1.0.post1
openai==1.6.1
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.3.2
requests==2.32.3
pandas==2.0.3
openpyxl==3.1.2 